% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bindDefectNumber.R
\name{bindDefectNumber}
\alias{bindDefectNumber}
\title{Create a map with a specified number of faults}
\usage{
bindDefectNumber(probabilityMatrix, faultValue, notFaultValue, faultNumber)
}
\arguments{
\item{probabilityMatrix:}{the matrix which represent the probability of every chip to be faulty}

\item{faultValue:}{the value which identifies a faulty chip}

\item{notFaultValue:}{the value which identifies a good (not faulty) chip}

\item{faultNumber:}{the number of fault in the matrix which is returned}
}
\value{
a matrix where exactly faultNumber elements are set to faultValue.
}
\description{
Accept a matrix which is assumed to represent the probability a single chip
on the wafer is to be faulty. The fuction returns the a matrix of the same
dimension of the passed one matrix where only the specified number of element
is set to the value which represent a fauly chip.
In other words, this function samples a subset of the chips to be faulty
and saves the other.
}

