% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gaussianDensity.R
\name{gaussianDensity}
\alias{gaussianDensity}
\title{Return a bivariate gaussian density}
\usage{
gaussianDensity(axes, mu, sigma)
}
\arguments{
\item{axes:}{a list of 2 elements: the first is assumed to be the x axes
and the second the y axes where the function should be evaluated}

\item{mu:}{a vector containing the mean of the gaussian density}

\item{sigma:}{a matrix containing the variance of the gaussian density}
}
\value{
a list with 2 elemets: the values of the probability function "pdf" and the
grid "grid" where it has been computed.
}
\description{
Computes a bivariate gaussian probability density and return a list containing
the grid used to compute the probability and a matrix with the values of the
probability function.
}

