% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parabolicDensity.R
\name{parabolicDensity}
\alias{parabolicDensity}
\title{Return a parabolic density}
\usage{
parabolicDensity(coefficient, ray, axes)
}
\arguments{
\item{axes:}{a list of 2 elements: the first is assumed to be the x axes
and the second the y axes where the function should be evaluated}

\item{coefficient:}{the coefficient a in the formula z = a(x^2 + y^2)}

\item{ray:}{the ray of the circular wafer}
}
\value{
list with 2 elemets: the values of the probability function "pdf" and the
grid "grid" where it has been computed.
}
\description{
Computes a parabolic probability density and return a list containing
the grid used to compute the probability and a matrix with the values of the
probability function. The center of the paraboloid is the one of the wafer.
}

