% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uniformDensity.R
\name{uniformDensity}
\alias{uniformDensity}
\title{Return a uniform density}
\usage{
uniformDensity(axes)
}
\arguments{
\item{axes:}{a list of 2 elements: the first is assumed to be the x axes
and the second the y axes where the function should be evaluated}
}
\value{
a list with 2 elemets: the values of the probability function "pdf" and the
grid "grid" where it has been computed.
}
\description{
Computes a uniform probability density and return a list containing
the grid used to compute the probability and a matrix with the values of the
probability function.
}

